added table of currencies rates
