+++ converter.html
@@ -6,6 +6,7 @@
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 		<title>Konwerter walut</title>
 		<link rel="stylesheet" href="./scss/main.css" />
+		<script src="./js/converter/currencyConverter.js" defer type="module"></script>
 	</head>
 	<body>
 		<nav>
@@ -19,6 +20,5 @@
 			<input type="button" value="Calculate" id="calcButton" />
 			<div id="calcResult" class="calcResult"></div>
 		</div>
-		<script src="./js/converter/currencyConverter.js"></script>
 	</body>
 </html>
