+++ index.html
@@ -6,14 +6,21 @@
 		<meta name="viewport" content="width=device-width, initial-scale=1.0" />
 		<title>Kursy walut</title>
 		<link rel="stylesheet" href="./scss/main.css" />
+		<script src="./js/charts/currencyCharts.js" defer type="module"></script>
+		<script src="https://cdn.jsdelivr.net/npm/chart.js@3.7.1/dist/chart.min.js"></script>
 	</head>
 	<body>
 		<nav>
 			<a href="converter.html">Konwerter walut</a>
 			<a href="index.html">Kursy walut</a>
 		</nav>
-		<h1>Wykresy</h1>
-
-		<script src="./js/charts/currencyCharts.js"></script>
+		<h1>Kursy walut</h1>
+		<h2>W stosunku do PLN</h2>
+		<table class="currencyTable">
+			<tr>
+				<th>Waluta</th>
+				<th>Wartość</th>
+			</tr>
+		</table>
 	</body>
 </html>
