+++ js/converter/currencyConverter.js
@@ -1,12 +1,4 @@
-const baseUrl = "api.frankfurter.app";
-
-async function getCurrencies() {
-	return new Promise(resolve => {
-		fetch(`https://${baseUrl}/currencies`)
-			.then(resp => resp.json())
-			.then(data => resolve(data));
-	});
-}
+import { getCurrencies, getAllRates } from "../apiServices.js";
 
 async function createSelect() {
 	const currencies = await getCurrencies();
@@ -18,23 +10,6 @@ async function createSelect() {
 		document.querySelector("#currenciesSelectFirst").appendChild(newCurrency1);
 		document.querySelector("#currenciesSelectSecond").appendChild(newCurrency2);
 	}
-}
-
-async function getAllRates() {
-	const currencies = await getCurrencies();
-	let currenciesUrl = `https://${baseUrl}/latest?to=`;
-	const currenciesLength = Object.keys(currencies).length;
-	let index = 0;
-	for (const key in currencies) {
-		if (currenciesLength - 1 == index) currenciesUrl += `${key}`;
-		else currenciesUrl += `${key},`;
-		index++;
-	}
-	return new Promise(resolve => {
-		fetch(currenciesUrl)
-			.then(resp => resp.json())
-			.then(data => resolve(data.rates));
-	});
 }
 
 createSelect();
